refactoring controller classes names create PostRest, PostService

+++ src/main/java/com/rubincomputers/sb_demo01/web/security/SecurityUtil.java
@@ -10,7 +10,7 @@ public class SecurityUtil {
     private SecurityUtil() {
     }
 
-    public static int authUserId() {
+    public static long authUserId() {
         return id;
     }
